chore: refacto on sign in form + changed prisma conf
