+++ prisma/schema.prisma
@@ -5,6 +5,7 @@ datasource db {
  
 generator client {
   provider        = "prisma-client-js"
+  binaryTargets = ["native", "rhel-openssl-3.0.x"]
 }
 
 model Account {
