+++ src/components/custom/sign-in/sign-in-form.tsx
@@ -1,0 +1,59 @@
+"use client"
+
+import { connectionSchema } from "@/validators/schemas/connectionSchema"
+import { UseFormReturn } from "react-hook-form"
+import { z } from "zod"
+
+import { Button } from "@/components/ui/button"
+import {
+  Form,
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage
+} from "@/components/ui/form"
+import { Input } from "@/components/ui/input"
+
+export default function SignInForm({ ...props }: SignInFormProps) {
+  return (
+    <Form {...props.form}>
+      <form onSubmit={props.form.handleSubmit(props.onSubmit)} className="space-y-8">
+        <FormField
+          control={props.form.control}
+          name="email"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Mail</FormLabel>
+              <FormControl>
+                <Input {...field} />
+              </FormControl>
+              <FormDescription>This is the email you used to register.</FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <FormField
+          control={props.form.control}
+          name="password"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Password</FormLabel>
+              <FormControl>
+                <Input {...field} type="password" />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <Button type="submit">Submit</Button>
+      </form>
+    </Form>
+  )
+}
+
+type SignInFormProps = {
+  form: UseFormReturn<z.infer<typeof connectionSchema>>
+  onSubmit: (values: z.infer<typeof connectionSchema>) => void
+}
