+++ src/lib/auth.ts
@@ -45,6 +45,5 @@ export const { handlers, auth, signIn, signOut } = NextAuth({
   },
   session: {
     strategy: "jwt"
-  },
-  trustHost: true
+  }
 })
